// frontend/src/app/api/models/destino-detalle.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ContenidoRelacionado } from '../models/contenido-relacionado';
import { Dificultad } from '../models/dificultad';
import { FuentePublica } from '../models/fuente-publica';
import { GuiaTarjeta } from '../models/guia-tarjeta';
import { ImagenPublica } from '../models/imagen-publica';
import { ItinerarioTarjeta } from '../models/itinerario-tarjeta';
import { Mes } from '../models/mes';
import { MetadatosEditoriales } from '../models/metadatos-editoriales';
import { PaisPublico } from '../models/pais-publico';
import { Presupuesto } from '../models/presupuesto';
import { RefTaxonomia } from '../models/ref-taxonomia';
import { Slug } from '../models/slug';
import { TerminoRef } from '../models/termino-ref';
import { TextoEnriquecido } from '../models/texto-enriquecido';
export interface DestinoDetalle {
  altitud_max_m?: (number | null);
  clima: string;
  como_llegar: TextoEnriquecido;
  descripcion_experta: TextoEnriquecido;
  dificultad: Dificultad;
  duracion_max_dias: number;
  duracion_min_dias: number;
  fuentes: Array<FuentePublica>;
  galeria: Array<ImagenPublica>;
  guias_relacionadas: Array<GuiaTarjeta>;
  itinerarios: Array<ItinerarioTarjeta>;

  /**
   * Solo si el destino no tiene itinerarios (SCR-004).
   */
  itinerarios_afines: Array<ItinerarioTarjeta>;
  itinerarios_total: number;
  latitud: number;
  longitud: number;
  meses_mejor_epoca: Array<Mes>;
  metadatos: MetadatosEditoriales;
  nivel_presupuesto: Presupuesto;
  pais: PaisPublico;
  portada: ImagenPublica;
  relacionados: Array<ContenidoRelacionado>;
  resumen: string;
  seguridad_riesgos: TextoEnriquecido;
  slug: Slug;
  sostenibilidad: TextoEnriquecido;
  terminos: Array<TerminoRef>;
  tipo_principal: RefTaxonomia;
  tipos: Array<RefTaxonomia>;
  titulo: string;
}