// frontend/src/app/api/models/destino-mapa-item.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Dificultad } from '../models/dificultad';
import { RefTaxonomia } from '../models/ref-taxonomia';
import { Slug } from '../models/slug';
export interface DestinoMapaItem {
  dificultad: Dificultad;
  latitud: number;
  longitud: number;
  pais: RefTaxonomia;
  region: RefTaxonomia;
  slug: Slug;
  titulo: string;
}