// frontend/src/app/api/models/destino-tarjeta.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Dificultad } from '../models/dificultad';
import { ImagenPublicaNula } from '../models/imagen-publica-nula';
import { RefTaxonomia } from '../models/ref-taxonomia';
import { Slug } from '../models/slug';
export interface DestinoTarjeta {
  dificultad: Dificultad;
  pais: RefTaxonomia;
  portada: ImagenPublicaNula;
  region: RefTaxonomia;
  slug: Slug;
  tipos: Array<RefTaxonomia>;
  titulo: string;
}