// frontend/src/app/api/models/guia-tarjeta.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ImagenPublicaNula } from '../models/imagen-publica-nula';
import { RefTaxonomia } from '../models/ref-taxonomia';
import { Slug } from '../models/slug';
export interface GuiaTarjeta {
  categoria: RefTaxonomia;
  fecha_ultima_revision: string;

  /**
   * ceil(palabras/200) (EXP-004, COULD).
   */
  minutos_lectura?: (number | null);
  portada: ImagenPublicaNula;
  resumen: string;
  slug: Slug;
  titulo: string;
}